Request confirmation of hte target surface in the output stream
Make an IPC transaction to the fraemwork to check if it wants to change the
rendering target. If it does, switch out the target or keep it the same in the
event no changes are required.

This completes libcameraservice portion of the "Report the surface in a camera
output stream" task in
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/19. It does not
include the changes in frameworks/base.

Change-Id: I0476901fde70b29937c4ce85488df60f027f44c9